wrap `cacheKey` usages with zero length check

+++ src/cpp/session/modules/SessionDataViewer.R
@@ -584,24 +584,25 @@
       }
    }
    
-   # if the object exists in the cache environment, return it. objects
-   # in the cache environment have already been coerced to data frames.
-   if (exists(cacheKey, where = .rs.CachedDataEnv, inherits = FALSE))
-      return(get(cacheKey, envir = .rs.CachedDataEnv, inherits = FALSE))
-   
-   # perhaps the object has been saved? attempt to load it into the
-   # cached environment
-   cacheFile <- file.path(cacheDir, paste(cacheKey, "Rdata", sep = "."))
-   if (file.exists(cacheFile))
-   {
-      status <- try(load(cacheFile, envir = .rs.CachedDataEnv), silent = TRUE)
-      if (inherits(status, "try-error"))
-         return(NULL)
-      
+   if (!.rs.isZeroLength(cacheKey)) {
+      # if the object exists in the cache environment, return it. objects
+      # in the cache environment have already been coerced to data frames.
       if (exists(cacheKey, where = .rs.CachedDataEnv, inherits = FALSE))
          return(get(cacheKey, envir = .rs.CachedDataEnv, inherits = FALSE))
+      
+      # perhaps the object has been saved? attempt to load it into the
+      # cached environment
+      cacheFile <- file.path(cacheDir, paste(cacheKey, "Rdata", sep = "."))
+      if (file.exists(cacheFile))
+      {
+         status <- try(load(cacheFile, envir = .rs.CachedDataEnv), silent = TRUE)
+         if (inherits(status, "try-error"))
+            return(NULL)
+         
+         if (exists(cacheKey, where = .rs.CachedDataEnv, inherits = FALSE))
+            return(get(cacheKey, envir = .rs.CachedDataEnv, inherits = FALSE))
+      }
    }
-   
    # failure
    return(NULL)
 })
@@ -752,6 +753,9 @@
    # save a copy into the cached environment
    cacheKey <- .rs.addCachedData(force(x), name)
    
+   if (.rs.isZeroLength(cacheKey))
+      return(invisible(NULL))
+
    # call viewData 
    invisible(.Call("rs_viewData", x, expr, title, name, env, cacheKey, FALSE))
 })
@@ -782,7 +786,8 @@
 
 .rs.addFunction("viewDataFrame", function(x, title, preview) {
    cacheKey <- .rs.addCachedData(force(x), "")
-   invisible(.Call("rs_viewData", x, "", title, "", emptyenv(), cacheKey, preview))
+   if (!.rs.isZeroLength(cacheKey))
+      invisible(.Call("rs_viewData", x, "", title, "", emptyenv(), cacheKey, preview))
 })
 
 .rs.addFunction("initializeDataViewer", function(server) {
@@ -807,7 +812,8 @@
 {
    # coerce to data frame before assigning, and don't assign if we can't coerce
    frame <- .rs.toDataFrame(obj, objName, TRUE)
-   if (!is.null(frame))
+   if (!is.null(frame) &&
+       !.rs.isZeroLength(cacheKey))
       assign(cacheKey, frame, .rs.CachedDataEnv)
 })
 
@@ -817,19 +823,20 @@
    if (Encoding(cacheDir) == "unknown")
       Encoding(cacheDir) <- "UTF-8"
    
-   # remove data from the cache environment
-   if (exists(".rs.CachedDataEnv") &&
-       exists(cacheKey, where = .rs.CachedDataEnv, inherits = FALSE))
-      rm(list = cacheKey, envir = .rs.CachedDataEnv, inherits = FALSE)
-   
-   # remove data from the cache directory
-   cacheFile <- file.path(cacheDir, paste(cacheKey, "Rdata", sep = "."))
-   if (file.exists(cacheFile))
-      file.remove(cacheFile)
-   
-   # remove any working data
-   .rs.removeWorkingData(cacheKey)
-   
+   if (!.rs.isZeroLength(cacheKey)) {
+      # remove data from the cache environment
+      if (exists(".rs.CachedDataEnv") &&
+         exists(cacheKey, where = .rs.CachedDataEnv, inherits = FALSE))
+         rm(list = cacheKey, envir = .rs.CachedDataEnv, inherits = FALSE)
+      
+      # remove data from the cache directory
+      cacheFile <- file.path(cacheDir, paste(cacheKey, "Rdata", sep = "."))
+      if (file.exists(cacheFile))
+         file.remove(cacheFile)
+      
+      # remove any working data
+      .rs.removeWorkingData(cacheKey)
+   }
    invisible(NULL)
 })
 
@@ -845,9 +852,10 @@
    
    # save each active cache file from the cache environment
    lapply(ls(.rs.CachedDataEnv), function(cacheKey) {
-      save(list = cacheKey, 
-           file = file.path(cacheDir, paste(cacheKey, "Rdata", sep = ".")),
-           envir = .rs.CachedDataEnv)
+      if (!.rs.isZeroLength(cacheKey))
+         save(list = cacheKey, 
+            file = file.path(cacheDir, paste(cacheKey, "Rdata", sep = ".")),
+            envir = .rs.CachedDataEnv)
    })
    
    # clean the cache environment
@@ -860,7 +868,8 @@
 
 .rs.addFunction("findWorkingData", function(cacheKey)
 {
-   if (exists(".rs.WorkingDataEnv") &&
+   if (!.rs.isZeroLength(cacheKey) &&
+       exists(".rs.WorkingDataEnv") &&
        exists(cacheKey, where = .rs.WorkingDataEnv, inherits = FALSE))
       get(cacheKey, envir = .rs.WorkingDataEnv, inherits = FALSE)
    else
@@ -869,7 +878,8 @@
 
 .rs.addFunction("removeWorkingData", function(cacheKey)
 {
-   if (exists(".rs.WorkingDataEnv") &&
+   if (!.rs.isZeroLength(cacheKey) &&
+       exists(".rs.WorkingDataEnv") &&
        exists(cacheKey, where = .rs.WorkingDataEnv, inherits = FALSE))
       rm(list = cacheKey, envir = .rs.WorkingDataEnv, inherits = FALSE)
    invisible(NULL)
@@ -877,7 +887,8 @@
 
 .rs.addFunction("assignWorkingData", function(cacheKey, obj)
 {
-   assign(cacheKey, obj, .rs.WorkingDataEnv)
+   if (!.rs.isZeroLength(cacheKey))
+      assign(cacheKey, obj, .rs.WorkingDataEnv)
 })
 
 .rs.addFunction("findGlobalData", function(name)
@@ -888,5 +899,10 @@
          return(name)
    }
    invisible("")
+})
+
+.rs.addFunction("isZeroLength", function(x)
+{
+   (x == "") || is.null(x) || is.na(x)
 })
